New function for making definition for manually unhandled input. It is possible using Mathematica to specify the input of functions in absolutely concrete way. The bad side of this is that for all functions one has to write func[args___] case, which leads to terminating of mathematica session (or some other behavior). Plus, one would like to get some relevant output before termination telling about details of an error. It is possible to do this manually for all functions, but it requieres a lot of time and slows down code implementation and leads to wish of writing a code in a "cowboy" style.
This is why I suggest to add and use MakeUnknownInputDefinition function (the current name is just a suggestion) to
speed this up and make Mathematica part of evaluation safer.

It should be used in the following way:
myFunc::usage = "text";
more messages
myFunc[f_] := some lines of code;
myFunc[some patterns] := some lines of code;
more definitions for myFunc

Utils`MakeUnknownInputDefinition[myFunc];

The last line creates a new definition for myFunc:
myFunc[args___]
which is evaluated if none of manually specified cases can be applied.

If this last definition is evaluated, then error message is generated and in contains information about:
myFunc::usage (if it is defined),
definitions of myFunc specified manually,
evaluated form of args.