fixing CMSSM librarylink SLHA test
Note 1: The librarylink prints all spectrum generator settings to SLHA
format.  The command line spectrum generator only prints the settings
that have been given in the input.  We may change this in the future,
so that the command line spectrum generator also prints all settings
that have been used.  This can be done by calling:

     slha_io.set_settings(settings);

where settings are the settings of the spectrum generator that are
used.